requirements.txt
# --- AI & RAG Core ---
llama-index
llama-index-llms-ollama
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
chromadb
llama-parse
rank_bm25
sentence-transformers
nltk
nest_asyncio

# --- Backend API ---
fastapi
uvicorn
pydantic
python-multipart
python-dotenv